+++ src/Core/Cartridge/Cartridge.h
@@ -3,6 +3,7 @@
 #include <vector>
 #include <string>
 #include <memory>
+#include "Common/Common.h"
 #include "Core/Cartridge/Mappers/Mapper.h"
 #include "Core/Cartridge/ROM/PRGROM.h"
 #include "Core/Cartridge/ROM/CHRROM.h"
@@ -26,11 +27,14 @@ namespace R2NES::Core
         // Retorna se o cartucho foi carregado com sucesso
         bool isValid() const { return imageValid; }
 
+        MirrorMode getMirrorMode() const { return mirror; }
+
     private:
         std::unique_ptr<PRGROM> prgROM;
         std::unique_ptr<CHRROM> chrROM;
         std::shared_ptr<Mapper> pMapper;
 
+        MirrorMode mirror = MirrorMode::HORIZONTAL;
         uint8_t mapperID = 0;
         uint8_t prgBanks = 0;
         uint8_t chrBanks = 0;